pyproject.toml
[project]
name = "market-research-gpt"
version = "0.1.0"
description = "Multi-agent market research assistant powered by LangGraph and Streamlit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "pdfplumber>=0.11.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
market-research-gpt = "app:main"

[tool.setuptools.packages.find]
include = ["agents*", "utils*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
